+++ lib/stdlib.h
@@ -1,9 +1,11 @@
 #ifndef __STDLIB_H
 #define __STDLIB_H
 
+#include "sys/types.h"
+
 int atoi(const char *);
 
-void *malloc(unsigned);
+void *malloc(size_t);
 void free(void *);
 
 void exit(int);
